Added UI Interdace and a few functionalities
Here's the list of what I've added so far:

Designed a StartScreen form as the initial screen for the game.

Designed buttons for "Play," "Exit," "Settings," and "About" in the StartScreen form.

Set button styles to have rounded corners and removed border lines.

Implemented functionality to navigate to the main game form on button click.

Added background music (bgForm.wav) to the StartScreen form, which plays on loop.

Designed a SettingsForm to configure game settings.

Added a switch to enable or disable background music in the SettingsForm.

Implemented a click sound feature for buttons, which plays a sound (clickSound.wav) when clicking any button.

Added the ability to toggle the click sound on and off in the SettingsForm with custom Swtich Controls.

Designed the StartScreen and SettingsForm forms to have custom minimize, maximize, and close buttons with custom icons.

Implemented logic to maintain the state of background music when switching between forms, preventing music from restarting.

This to do/debug:
Address the issue where the background music continues to play even when the toggle is off in the SettingsForm.

Fix the click sound feature to ensure it plays when clicking buttons.

Resolve the issue with the game music switch not working.

Things Missing: The actual game functionality, as of now I added a picture box where the game should display